requirements.txt
# Core dependencies
typer==0.15.4
python-dotenv>=1.0.0
pyyaml==6.0.2
requests==2.32.3

# MAAS API client
python-libmaas>=0.6.8

# SSH and remote execution
paramiko==3.5.1
bcrypt>=4.1.2

# Kubernetes
kubernetes==32.0.1

# Testing
pytest==8.3.5
pytest-asyncio>=0.23.0

# Data validation
jsonschema>=4.24.0
jsonschema-specifications>=2023.3.6
referencing>=0.28.4
rpds-py>=0.7.1

# Async support
aiohttp>=3.9.0
asyncio>=3.4.3

# Security-related
dnspython>=2.0.0
PyNaCl>=1.5.0
macaroonbakery>=1.1.3

setuptools

# Templating
Jinja2>=3.1.4
